services/api/pyproject.toml
[project]
name = "nsos-cpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.8",
    "django>=6.0",
    "django-allauth[headless]>=65.14.0",
    "django-cors-headers>=4.9.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "sqlparse>=0.5.5",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "boto3-stubs>=1.42.8",
    "django-stubs[compatible-mypy]>=5.2.9",
    "mypy-boto3-s3>=1.42.21",
    "mypy-boto3-secretsmanager>=1.42.8",
    "mypy-boto3-ses>=1.42.3",
    "typing-extensions>=4.15.0",
]
